Refactored Driver
I refactored the driver to reuse more code. Now instead of each click handler basically being a mini-program, each handler just calls the problemDriver() function.

ProblemDriver() accepts a problem name as input and uses that problem name to access HTML elements using JQuery and uses it as a key to find the appropriate implementation inside of a dispatch table.